fix(stella-tui): move the eviction rebase into its own module
Three gates were red on this branch.

The file-size ratchet: the rebase logic put `deck_ui.rs` twenty-two lines over
its ceiling. The guard's own remedy is a sibling module, so the loop moves to
`deck_ui/eviction.rs` and `clamp` calls it. `deck_ui.rs` drops to 3660 lines,
twenty-eight under what main has.

Formatting: `transcript_nav.rs` was unformatted, and the shape rustfmt chose
for the assertion then tripped `clippy::possible_missing_comma`. The
subtraction binds to a local, which both tools accept.

The new module's comments are written to the reading-grade ceiling a new file
is held to.

Refs #5371
